Quick part picker refactor VAB now has a VehicleRoot object to pul all of the parts under
