+++ Assets/Scripts/VAB/PartPickerPanel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cf21e67f754067c4ca4bbe59398041e1
